requirements.txt
pyDes
# sasl
# thrift
# thrift-sasl
# PyHive
hdfs
pyspark
apache-airflow-providers-apache-spark
apache-airflow-providers-apache-hive
psycopg2